HSEARCH-3105 Split the inner classes in their own files
