+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/dsl/impl/IntegerIndexSchemaFieldContext.java
@@ -6,29 +6,15 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.dsl.impl;
 
-import java.util.Objects;
-
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.IntPoint;
-import org.apache.lucene.document.NumericDocValuesField;
-import org.apache.lucene.document.StoredField;
-import org.apache.lucene.index.IndexableField;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.search.SortField;
 import org.hibernate.search.v6poc.backend.document.impl.DeferredInitializationIndexFieldAccessor;
 import org.hibernate.search.v6poc.backend.document.model.Sortable;
-import org.hibernate.search.v6poc.backend.document.model.Store;
-import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneIndexFieldAccessor;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldFormatter;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldQueryFactory;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaFieldNode;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaNodeCollector;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.MatchQueryOptions;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.RangeQueryOptions;
-import org.hibernate.search.v6poc.backend.lucene.search.sort.impl.LuceneSearchSortCollector;
-import org.hibernate.search.v6poc.search.dsl.sort.SortOrder;
+import org.hibernate.search.v6poc.backend.lucene.types.formatter.impl.IntegerFieldFormatter;
+import org.hibernate.search.v6poc.backend.lucene.types.predicate.impl.IntegerFieldQueryFactory;
+import org.hibernate.search.v6poc.backend.lucene.types.sort.impl.IntegerFieldSortContributor;
 
 /**
  * @author Guillaume Smet
@@ -61,129 +47,5 @@ public class IntegerIndexSchemaFieldContext extends AbstractLuceneIndexSchemaFie
 		accessor.initialize( new LuceneIndexFieldAccessor<>( schemaNode ) );
 
 		collector.collectFieldNode( schemaNode.getAbsoluteFieldPath(), schemaNode );
-	}
-
-	private static final class IntegerFieldFormatter implements LuceneFieldFormatter<Integer> {
-
-		private final Store store;
-
-		private final Sortable sortable;
-
-		private IntegerFieldFormatter(Store store, Sortable sortable) {
-			this.store = store;
-			this.sortable = sortable;
-		}
-
-		@Override
-		public void addFields(LuceneDocumentBuilder documentBuilder, LuceneIndexSchemaObjectNode parentNode, String fieldName, Integer value) {
-			if ( value == null ) {
-				return;
-			}
-
-			if ( Store.YES.equals( store ) ) {
-				documentBuilder.addField( parentNode, new StoredField( fieldName, value ) );
-			}
-
-			if ( Sortable.YES.equals( sortable ) ) {
-				documentBuilder.addField( parentNode, new NumericDocValuesField( fieldName, value.longValue() ) );
-			}
-
-			documentBuilder.addField( parentNode, new IntPoint( fieldName, value ) );
-		}
-
-		@Override
-		public Integer parse(Document document, String fieldName) {
-			IndexableField field = document.getField( fieldName );
-
-			if ( field == null ) {
-				return null;
-			}
-
-			return (Integer) field.numericValue();
-		}
-
-		@Override
-		public Object format(Object value) {
-			return value;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( this == obj ) {
-				return true;
-			}
-			if ( obj == null ) {
-				return false;
-			}
-			if ( IntegerFieldFormatter.class != obj.getClass() ) {
-				return false;
-			}
-
-			IntegerFieldFormatter other = (IntegerFieldFormatter) obj;
-
-			return Objects.equals( store, other.store ) && Objects.equals( sortable, other.sortable );
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash( store, sortable );
-		}
-	}
-
-	private static final class IntegerFieldQueryFactory implements LuceneFieldQueryFactory {
-
-		private static final IntegerFieldQueryFactory INSTANCE = new IntegerFieldQueryFactory();
-
-		private IntegerFieldQueryFactory() {
-		}
-
-		@Override
-		public Query createMatchQuery(String fieldName, Object value, MatchQueryOptions matchQueryOptions) {
-			return IntPoint.newExactQuery( fieldName, (Integer) value );
-		}
-
-		@Override
-		public Query createRangeQuery(String fieldName, Object lowerLimit, Object upperLimit, RangeQueryOptions rangeQueryOptions) {
-			return IntPoint.newRangeQuery(
-					fieldName,
-					getLowerValue( lowerLimit, rangeQueryOptions.isExcludeLowerLimit() ),
-					getUpperValue( upperLimit, rangeQueryOptions.isExcludeUpperLimit() )
-			);
-		}
-
-		private static int getLowerValue(Object lowerLimit, boolean excludeLowerLimit) {
-			if ( lowerLimit == null ) {
-				return Integer.MIN_VALUE;
-			}
-			else {
-				return excludeLowerLimit ? Math.addExact( (int) lowerLimit, 1 ) : (int) lowerLimit;
-			}
-		}
-
-		private static int getUpperValue(Object upperLimit, boolean excludeUpperLimit) {
-			if ( upperLimit == null ) {
-				return Integer.MAX_VALUE;
-			}
-			else {
-				return excludeUpperLimit ? Math.addExact( (int) upperLimit, -1 ) : (int) upperLimit;
-			}
-		}
-	}
-
-	private static class IntegerFieldSortContributor extends AbstractScalarLuceneFieldSortContributor {
-
-		private static final IntegerFieldSortContributor INSTANCE = new IntegerFieldSortContributor();
-
-		private IntegerFieldSortContributor() {
-			super( Integer.MIN_VALUE, Integer.MAX_VALUE );
-		}
-
-		@Override
-		public void contribute(LuceneSearchSortCollector collector, String absoluteFieldPath, SortOrder order, Object missingValue) {
-			SortField sortField = new SortField( absoluteFieldPath, SortField.Type.INT, order == SortOrder.DESC ? true : false );
-			setEffectiveMissingValue( sortField, missingValue, order );
-
-			collector.collectSortField( sortField );
-		}
 	}
 }
