+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/dsl/impl/StringIndexSchemaFieldContext.java
@@ -6,38 +6,22 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.dsl.impl;
 
-import java.util.Objects;
-
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.Field;
 import org.apache.lucene.document.FieldType;
-import org.apache.lucene.document.SortedDocValuesField;
 import org.apache.lucene.index.IndexOptions;
-import org.apache.lucene.index.Term;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.search.SortField;
-import org.apache.lucene.search.TermQuery;
-import org.apache.lucene.search.TermRangeQuery;
-import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.QueryBuilder;
 import org.hibernate.search.v6poc.backend.document.impl.DeferredInitializationIndexFieldAccessor;
 import org.hibernate.search.v6poc.backend.document.model.IndexSchemaFieldTypedContext;
 import org.hibernate.search.v6poc.backend.document.model.Sortable;
 import org.hibernate.search.v6poc.backend.document.model.Store;
-import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneIndexFieldAccessor;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldFormatter;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldQueryFactory;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaFieldNode;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaNodeCollector;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.MatchQueryOptions;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.RangeQueryOptions;
-import org.hibernate.search.v6poc.backend.lucene.search.sort.impl.LuceneSearchSortCollector;
-import org.hibernate.search.v6poc.backend.lucene.util.impl.AnalyzerUtils;
-import org.hibernate.search.v6poc.search.dsl.sort.SortOrder;
+import org.hibernate.search.v6poc.backend.lucene.types.formatter.impl.StringFieldFormatter;
+import org.hibernate.search.v6poc.backend.lucene.types.predicate.impl.StringFieldQueryFactory;
+import org.hibernate.search.v6poc.backend.lucene.types.sort.impl.StringFieldSortContributor;
 
 /**
  * @author Guillaume Smet
@@ -139,165 +123,5 @@ public class StringIndexSchemaFieldContext extends AbstractLuceneIndexSchemaFiel
 		fieldType.freeze();
 
 		return fieldType;
-	}
-
-	private static final class StringFieldFormatter implements LuceneFieldFormatter<String> {
-
-		private final Sortable sortable;
-
-		private final FieldType fieldType;
-
-		private final Analyzer normalizer;
-
-		private StringFieldFormatter(Sortable sortable, FieldType fieldType, Analyzer normalizer) {
-			this.sortable = sortable;
-			this.fieldType = fieldType;
-			this.normalizer = normalizer;
-		}
-
-		@Override
-		public void addFields(LuceneDocumentBuilder documentBuilder, LuceneIndexSchemaObjectNode parentNode, String fieldName, String value) {
-			if ( value == null ) {
-				return;
-			}
-
-			documentBuilder.addField( parentNode, new Field( fieldName, value, fieldType ) );
-
-			if ( Sortable.YES.equals( sortable ) ) {
-				documentBuilder.addField( parentNode, new SortedDocValuesField(
-						fieldName,
-						new BytesRef( normalizer != null ? AnalyzerUtils.analyzeSortableValue( normalizer, fieldName, value ) : value )
-				) );
-			}
-		}
-
-		@Override
-		public Object format(Object value) {
-			return value;
-		}
-
-		@Override
-		public String parse(Document document, String fieldName) {
-			return document.get( fieldName );
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( this == obj ) {
-				return true;
-			}
-			if ( obj == null ) {
-				return false;
-			}
-			if ( StringFieldFormatter.class != obj.getClass() ) {
-				return false;
-			}
-
-			StringFieldFormatter other = (StringFieldFormatter) obj;
-
-			return Objects.equals( sortable, other.sortable ) &&
-					Objects.equals( fieldType, other.fieldType ) &&
-					Objects.equals( normalizer, other.normalizer );
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash( sortable, fieldType, normalizer );
-		}
-	}
-
-	private static class StringFieldQueryFactory implements LuceneFieldQueryFactory {
-
-		private final Analyzer analyzerOrNormalizer;
-
-		private final boolean tokenized;
-
-		private final QueryBuilder queryBuilder;
-
-		private StringFieldQueryFactory(Analyzer analyzerOrNormalizer, boolean tokenized, QueryBuilder queryBuilder) {
-			this.analyzerOrNormalizer = analyzerOrNormalizer;
-			this.tokenized = tokenized;
-			this.queryBuilder = queryBuilder;
-		}
-
-		@Override
-		public Query createMatchQuery(String fieldName, Object value, MatchQueryOptions matchQueryOptions) {
-			String stringValue = (String) value;
-
-			if ( queryBuilder != null ) {
-				return queryBuilder.createBooleanQuery( fieldName, stringValue, matchQueryOptions.getOperator() );
-			}
-			else {
-				// we are in the case where we a have a normalizer here as the analyzer case has already been treated by
-				// the queryBuilder case above
-
-				return new TermQuery( new Term( fieldName, getAnalyzedValue( analyzerOrNormalizer, fieldName, stringValue ) ) );
-			}
-		}
-
-		@Override
-		public Query createRangeQuery(String fieldName, Object lowerLimit, Object upperLimit, RangeQueryOptions rangeQueryOptions) {
-			// Note that a range query only makes sense if only one token is returned by the analyzer
-			// and we should even consider forcing having a normalizer here, instead of supporting
-			// range queries on analyzed fields.
-
-			return TermRangeQuery.newStringRange(
-					fieldName,
-					getAnalyzedValue( analyzerOrNormalizer, fieldName, (String) lowerLimit ),
-					getAnalyzedValue( analyzerOrNormalizer, fieldName, (String) upperLimit ),
-					// we force the true value if the limit is null because of some Lucene checks down the hill
-					lowerLimit == null ? true : !rangeQueryOptions.isExcludeLowerLimit(),
-					upperLimit == null ? true : !rangeQueryOptions.isExcludeUpperLimit()
-			);
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( this == obj ) {
-				return true;
-			}
-			if ( obj == null ) {
-				return false;
-			}
-			if ( StringFieldFormatter.class != obj.getClass() ) {
-				return false;
-			}
-
-			StringFieldQueryFactory other = (StringFieldQueryFactory) obj;
-
-			return Objects.equals( analyzerOrNormalizer, other.analyzerOrNormalizer ) &&
-					tokenized == other.tokenized &&
-					Objects.equals( queryBuilder, other.queryBuilder );
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash( analyzerOrNormalizer, tokenized, queryBuilder );
-		}
-
-		private static String getAnalyzedValue(Analyzer analyzerOrNormalizer, String fieldName, String stringValue) {
-			if ( analyzerOrNormalizer == null ) {
-				return stringValue;
-			}
-
-			return AnalyzerUtils.analyzeSortableValue( analyzerOrNormalizer, fieldName, stringValue );
-		}
-	}
-
-	private static class StringFieldSortContributor extends AbstractScalarLuceneFieldSortContributor {
-
-		private static final StringFieldSortContributor INSTANCE = new StringFieldSortContributor();
-
-		private StringFieldSortContributor() {
-			super( SortField.STRING_FIRST, SortField.STRING_LAST );
-		}
-
-		@Override
-		public void contribute(LuceneSearchSortCollector collector, String absoluteFieldPath, SortOrder order, Object missingValue) {
-			SortField sortField = new SortField( absoluteFieldPath, SortField.Type.STRING, order == SortOrder.DESC ? true : false );
-			setEffectiveMissingValue( sortField, missingValue, order );
-
-			collector.collectSortField( sortField );
-		}
 	}
 }
